# apps/backend/src/openapi/v1/swagger_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BilboMD API Docs",
    "version": "1.21.1",
    "description": "API for external programmatic access to BilboMD job submission and results"
  },
  "servers": [
    {
      "url": "https://bilbomd.bl1231.als.lbl.gov/api/v1",
      "description": "Beamline 12.3.1"
    },
    {
      "url": "https://bilbomd-nersc.bl1231.als.lbl.gov/api/v1",
      "description": "NERSC"
    },
    {
      "url": "http://localhost:3501/api/v1",
      "description": "development"
    }
  ],
  "paths": {
    "/external/jobs": {
      "post": {
        "summary": "Submit a new job",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "oneOf": [
                  {
                    "title": "PDB Job",
                    "required": [
                      "title",
                      "bilbomd_mode",
                      "pdb_file",
                      "crd_file",
                      "const_inp_file",
                      "dat_file"
                    ],
                    "properties": {
                      "bilbomd_mode": {
                        "type": "string",
                        "enum": [
                          "pdb"
                        ],
                        "description": "Must be set to \"pdb\" for this job type."
                      },
                      "title": {
                        "type": "string",
                        "description": "User-defined title for the job."
                      },
                      "pdb_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the PDB file containing the atomic coordinates."
                      },
                      "crd_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the CHARMM CRD file associated with the structure."
                      },
                      "const_inp_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the const.inp configuration file used by BilboMD."
                      },
                      "dat_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the experimental SAXS .dat file for fitting analysis."
                      }
                    }
                  },
                  {
                    "title": "CRD/PSF Job",
                    "required": [
                      "title",
                      "bilbomd_mode",
                      "crd_file",
                      "psf_file",
                      "const_inp_file",
                      "dat_file"
                    ],
                    "properties": {
                      "bilbomd_mode": {
                        "type": "string",
                        "enum": [
                          "crd_psf"
                        ],
                        "description": "Must be set to \"crd_psf\" for this job type."
                      },
                      "title": {
                        "type": "string",
                        "description": "User-defined title for the job."
                      },
                      "crd_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the CHARMM CRD file containing structure coordinates."
                      },
                      "psf_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the PSF file containing topology information."
                      },
                      "const_inp_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the const.inp configuration file for BilboMD."
                      },
                      "dat_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the experimental SAXS .dat file for fitting."
                      }
                    }
                  },
                  {
                    "title": "Auto Job",
                    "required": [
                      "title",
                      "bilbomd_mode",
                      "pdb_file",
                      "pae_file",
                      "dat_file"
                    ],
                    "properties": {
                      "bilbomd_mode": {
                        "type": "string",
                        "enum": [
                          "auto"
                        ],
                        "description": "Must be set to \"auto\" for this job type."
                      },
                      "title": {
                        "type": "string",
                        "description": "User-defined title for the job."
                      },
                      "pdb_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the PDB file containing the full predicted structure."
                      },
                      "pae_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the PAE (Predicted Aligned Error) matrix file from AlphaFold."
                      },
                      "dat_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the experimental SAXS .dat file for fitting."
                      }
                    }
                  },
                  {
                    "title": "AlphaFold Job",
                    "required": [
                      "title",
                      "bilbomd_mode",
                      "dat_file",
                      "entities_json"
                    ],
                    "properties": {
                      "bilbomd_mode": {
                        "type": "string",
                        "enum": [
                          "alphafold"
                        ],
                        "description": "Must be set to \"alphafold\" for this job type."
                      },
                      "title": {
                        "type": "string",
                        "description": "User-defined title for the job."
                      },
                      "dat_file": {
                        "type": "string",
                        "format": "binary",
                        "description": "Upload the experimental SAXS .dat file for fitting."
                      },
                      "entities_json": {
                        "$ref": "#/components/schemas/EntitiesJson",
                        "description": "JSON file containing entity information for the job."
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "New auto job successfully created"
                    },
                    "jobid": {
                      "type": "string",
                      "example": "680aa4ffce4a325a086cbced"
                    },
                    "uuid": {
                      "type": "string",
                      "example": "f4ba7568-369a-465a-b0d1-a78969bb816b"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenAlphaFold"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntitiesJson": {
        "type": "object",
        "required": [
          "entities"
        ],
        "properties": {
          "entities": {
            "type": "array",
            "description": "List of entities to model, each with type, sequence, and number of copies.",
            "items": {
              "type": "object",
              "required": [
                "id",
                "type",
                "sequence",
                "copies"
              ],
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 1
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "Protein",
                    "DNA",
                    "RNA"
                  ],
                  "example": "Protein"
                },
                "sequence": {
                  "type": "string",
                  "example": "MSEQNNTEMTFQIQRIYTKDISFEAPNAPHVFQKDWLD..."
                },
                "copies": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "BilboMD API Token is missing or invalid",
        "content": {
          "application/json": {
            "examples": {
              "InvalidToken": {
                "summary": "Invalid API token",
                "value": {
                  "message": "Invalid API token"
                }
              },
              "MissingHeader": {
                "summary": "Missing or invalid Authorization header",
                "value": {
                  "message": "Missing or invalid Authorization header"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request failed validation",
        "content": {
          "application/json": {
            "example": {
              "message": "Validation failed",
              "errors": [
                {
                  "path": "dat_file",
                  "message": "No valid SAXS data found"
                },
                {
                  "path": "pae_file",
                  "message": "A PAE *.json file is required"
                },
                {
                  "path": "pae_file",
                  "message": "Only accepts a *.json file."
                },
                {
                  "path": "pae_file",
                  "message": "Max file size is 120MB"
                },
                {
                  "path": "pae_file",
                  "message": "Filename must be no longer than 30 characters."
                }
              ]
            }
          }
        }
      },
      "ForbiddenAlphaFold": {
        "description": "AlphaFold jobs are disabled due to configuration",
        "content": {
          "application/json": {
            "example": {
              "message": "AlphaFold jobs unavailable on this deployment."
            }
          }
        }
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}